Added a unit test module for the cli.
During the course of which I discovered that if the test module is in a
subdirectory of the source tree, it won't be able to import from the
root of the tree (We probably missed this before because we were
installing the haas into the venv). To remedy this, I moved the model
tests to the root of the tree, next to the ones for the cli.

Right now there's one test, which checks for proper error handling in
the case of an invald number of arguments passed to a command. It is
currently failing.