Slight readability enhancement in the README.
The long url in the middle of the text made this a bit awkward to read
(and it didn't really fit within the line-length the document is
shooting for). The footnote style links are better for this.

I'm going to add some links to testing documentation, wanted to clean
this up first.