tested the new model
One to one mapping is working. (every entity has a label, which is not a primary key, I don't know if we should ensure the label is unique within the same entity domain manually or just specify it as primary key. the one to one mapping is based on the labels, it will have subtle bugs if label is not unique).

Many-to-many mapping User<->Group is working.
The last few lines is a test for it.