Renamed the haas.api_server module to haas.server.
We all agreed on Friday that the name needed to change; I think this is
a reasonable solution.

One of the docstrings had a reference to api_server in a section that
was otherwise out of date. Rather than simply adjusting the name, I
updated the docstring to match the current situation.