Fix bad prototype for headnode_{create,delete}
In the API documentation, these functions take a headnode name and a /group/
name---but in api.py, they took a headnode name and a /project/ name.  This
doesn't make sense, because we also have the API call
'project_connect_headnode'.  This commit fixes both the function and its
corresponding unit test.