Fix Nic model
First, the node is NOT nullable, because Nics /always/ belong to nodes.  For
that reason, the node MUST also be part of the constructor.

Second, remove the 'group' field from Nics.  We argued that it was easier to
have everything explicitly list its group, but that is not true for Nics.
This is because they are closely attached to Nodes, and Nodes can change
groups!  This would require, on every group change of a node, searching for
all the Nics attached to it and changing their groups as well.

The purpose of this commit is to prepare for implementing node_register_nic
and node_delete_nic, which were accidentally skipped.