Wrote a helper function to generate api mappings with less boilerplate.
This will give us the ability to specify the rest mapping in a more
declarative way.

I'm not certain I think the form_params argument is really necessary; It
can be taken out without affecting functionality, but might serve to
make the code a bit more self-documenting.

I converted a couple of the http wrappers over to this, but haven't
done the rest of them yet; I'd like to get some feedback first.