Establish new driver API
The API is documented in haas/drivers/__init__.py .  There are two drivers
now: the dell driver and the null driver (which, as its name suggests, does
nothing.)

All unit tests are explicitly run against the null driver.  So, any
functionality, such as vlan allocation, that is specific to a drier, such as
the Dell driver, must get its own test suite.

There is one regression here: the Dell driver now has hard-coded which VLANs
it can use.  We should add a configuration option for that.

There is one HUGE piece of trickiness in the code: the init_db function calls
the driver's init_db---but table creation is done by the fact that we import
the module BEFORE doing table creation.  This is tricky, and confusing, and
terrible, and we should fix it.