Set the frozen bit when the VM is created.
In addition, we create the VM's nics inside it's own create method now,
rather than in project_deploy.

Before I go any further, I need to test this in a real environment -
I've touched code that isn't hit by the test suite because it needs a
real libvirt.

Still on the TODO list is making all of the relevant API calls check the
frozen bit.