Started tests for the headnode-freezing logic.
The one test I've written thus far is failing, since the relevant
functionality isn't actually implemented.

The `@no_dry_run` decorator around `Headnode.create` presents a problem
for this; we're setting the frozen bit inside of create, which means it
won't be set if we're just running the tests locally (with dry_run
enabled). Most likely the bit will have to be set in the api module
instead.