Check that HNIC is attached before attaching to bridge
Previously, the code would attempt to attach the HNIC to a network, and hit a
null dereference if it wasn't attached.  Now it simply skips creating the
HNIC.  It might be nice to create the HNIC but not attach it, but it's a
little harder to do that.

This fixes bug 195