Fix port_register bug
This patch fixes two problems.
In cli.py, quote by default assumes that '/' characters are
safe.  An empty string was provided to signify that is is
not safe and should be escaped.

In api.py, flask was decoding port names before evaluating a
URI.  This caused 404 errors when evaulating the URI if
there were '/' characters in the name.  To fix, 'path:' was
added before the port argument to signify that the argument
should not be decoded.