Console Feature Cleanup
This patch cleans up some of the code added in this feature branch.
Additionally, this patch also makes it so the user is in charge of starting and
stopping console logging via the API, instead of logging always happening while
a node belongs to a project.  This ensures that a large number of unused
logging subprocesses are not running.