Make arguments in the body json, not form data.
This change is transparent from the perspective of an API server, but
clients will have to be adjusted to send parameters in the new format.

The docstrings have been updated to reflect the change in format --
refer to them for more information.

This also adds a dependency on the schema library. This is overkill for
what we're using it for so far, but the plan is to allow the user to
supply custom schemas for their api functions.

This patch also adds a test to check arguments in the body.