Update README, write HACKING.
The README is now less developer-biased; It points to the INSTALL and
HACKING docs as appropriate.

This patch also updates a few things that were out of date, e.g. the
direct use of IPMI by users, and the mac address parameter to the
headnode_create_hnic call in one of the examples.