Allow caller of rest_call to provide a schema.
Without this, options for API design are severely limited.

Along the way I did a bit of cleanup. The most significant additional
change is the way the arguments to the API call are being marshalled to
from the request. It is semantically equivalent, but the implementation
is more straightforward.

Other changes include better error message, and moving a comment to a
more appropriate location.