Run each test in a separate process
This is noticeably slower, but saves us from having to try to support
"unloading" extensions, which is something we don't otherwise need, and
would be a pain.

This obviates the need for network_allocator.reset()

Also, this shook out a few other issues with the test suite, which have
been fixed.