Port serve_networks to new driver model
...and remove much of the now-obsolete code, including the top-level
network drivers from the old model.

Some of the tests for the old drivers are still present but marked
expected to fail --- they test functionality that we should be
exercising in the new driver model as well, so I want to keep them there
as a reference until the replacements are ready.