Implement most of the api-level changes
This is basically a cherry-pick of aae289b, except *done by hand*,
because the merge was *that* bad. This also avoids getting rid of
NetworkingAction, which was added back in later.

There are a few failing tests, but the next handful of patches are bug
fixes, so I'm going to cherry-pick those and see if fixes, probably
squashing after.