Fix the rest of the deployment test bitrot
The network-related deployment tests have been udpated to work with the
recent changes to the codebase.

Parts of the tests have been factored out into a superclass defined in
haas.test_common, as I have plans to add tests that exercise the vlan
trunking functionality, which will use some of the same techniques.