Dell driver: fix bug fetching attachment from DB
Apparently Class.query isn't something raw SQLAlchemy can do? I've done
this before in a flask project and it worked...

The current solution is less than great since it's actually pulling in
all attachments for the nic, but fussing about it is premature
optimization.

The right solution is to address #454.