Update example configuration files
The old example config files had bitrotted. They had lots of duplicated
comments, with only minor differences. This was deemed unsustainable,
and after talking to @henn, we decided a good solution was to have one
thoroughly documented file, and otherwise just keep the no-hardware dev
config, since that was an easy ready-to-run example, but just reference
the first example for descriptions of the options.

Thanks to @SahilTikale and @apoorvemohan for catching this.