Fix test failures introduced by merging master back in.
Git didn't raise any conflicts, but because @SahilTikale added mandatory
fields to the MockSwitch model, most of the auth tests were failing.

This patch isn't as big as it looks; in order to supply the arguments to
switch register in e.g. admin_calls, we needed to add a kwargs field, so
the diff includes many changes that are just adding `{}` to the end of a
tuple.