Fix bug caught by 9400f695 & remove Werkzeug dep.
In the course of the fix, I've simplified _make_schema_for such that it
no longer needs Werkzeug. This was the last place we were using Werkzeug
explicitly, so the dependency has been removed from setup.py. The
package will still be pulled in indirectly, since it is a dependency of
Flask.