Fix stress tests
All the non-deployment tests are now passing, but there's something
mysterious going on: we get an error in one of the tests in
`tests/unit/ext/auth/database.py` when all of `tests/unit` is run, but
not when just that file is run. Still need to figure that one out.