Remove wrapper fixtures around fresh_database
Tests used to take an explicit `db` parameter, so naming the test
fixture `db` caused it to be passed in to the tests. We no longer pass
it in explicitly, and in most places the wrapper didn't actually do
anything. The wrapper has been removed.

This has also fixed the mysterious test failures; it appears to have
been a naming conflict. There are some other things called `db` floating
around, and pytest is crap about namespacing.