Skeleton for haas-admin command
Flask-Migrate has some built-in support for the Flask-Script extension,
which does command line stuff. It'll be easier to work with this if we
don't have to glue it to `haas.cli`, and we've talked briefly about
splitting server/admin/client functionality into different commands
anyway.

Right now this is just an entry point in setup.py, with a function that
no-ops. Like I said, skeleton.