New constraints on the network allocator's interface.
The migration docs say to re-run `haas-admin db create` when adding new
extensions. For this to work it actually needs to be safe to do this,
including re-running the network allocator's populate() method.

The vlan_pool driver currently doesn't actually conform to these
constraints. A test has been added which detects this, and the desired
functionality has been added to the network-drivers doc.