pytest: Remove directory from addopts
This was why pytest was scanning `env.py`, so this also obviates our
`confest.py`. There's a commented out `addopts` directive that addes
code coverage; presumably whoever made that change didn't realize that
`haas` was an argument to the `--cov` option. Leaving this in there had
the affect of scanning the entire `haas` subdirectory for tests.