Remove user-aware code from core, add per-request IDs to logs.
The get_user method for auth backends was added because we needed a way
to correlate API calls with the credentials being used with a request.
However, this is an objectionable design, since the core doesn't
otherwise have a notion of user; this is something only the extensions
which actually implement users know about (not all of them necessarily
will).

Instead, we simply have the backends log this information themselves. In
order to be able to correlate the now separate log entries, we add a
request context ID to the logging calls. We introduce a generic
mechanism for request-specific logging.