Generalize the log_dir to affect the whole haas package
Per @knikola's request. The reason for the timing is that the previous
commit (moving the user logging to the extensions) has the effect of
moving logging statements that previously went to the haas.rest logger
elsewhere. After discussing on IRC, we decided there wasn't a good
reason to only send haas.rest here.