Fix pep8 error in migrations template
script.py.mako is used as a template for new migration scripts, but
previously did not conform to the pep8 tool's expectations. In
particular, the imports should have been at the top of the module.
script.py.mako is not itself checked by pep8 (and can't be, since the
templating language is not valid python), but the generated scripts are,
and it was generating scripts that would fail the pep8 checks.

The template has been modified to generate scripts closer to pep8
compliance, but note that alembic still does things like generate very
long lines, which can't be fixed in the template.