Fix bogus i18n call.
The `_` function is a common alias used when doing internationalization.
We don't actually do i18n yet, and there's no alias defined/imported in
this file, so if this statement gets executed, it raises a NameError.

This was introduced in commit 16016078453913918dcec208c2c000a7f935a488,
though it's not clear to me why.