Updated the Brocade driver to enable the switchport when setting mode
This step had to previously be done by the administrator through the
commandline interface by issuing the command ``switchport``. It
is now automated by the HIL driver by issuing an API call to
enable switching on the interface before the driver sets the port.

Also added logging to brocade driver if switch returns 4xx.