Fix #731
Here's a graph of the calls involved in this bug:

    HIL -+--> Flask-Migrate --------> Alembic ------+
         |                                          |
         +--> Flask-SQLAlchemy ---+-> SQLAlchemy  <-+
                                  |
              Flask <-------------+

    (Key: caller --> callee)

Previously, HIL called out to Flask-SQLAlchemy to create the
database schema, which resolves the database path relative
to the location where the Flask app is defined. This is
typical of flask; relative paths are relative to the app,
not the running process.

However, We called stamp() from Flask-Migrate to actually
add the revision identifiers. 'stamp()' just gets passed
right through to Alembic, which knows about SQLAlchemy but
crucially does *not* know about Flask. It interprets the same
path as being relative to the current working directory, and
so opens a DB there.

With this patch, we use Alembic to get the list of revision
identifiers, and insert them ourselves with Flask-SQLAlchemy.