Modified the dell and cisco drivers to fix the deployment tests. Collaborative effort with @zenhack
Both, the dell and cisco, drivers were timing out when running the tests.

In the dell driver, pexepct times out when we try to gather information for an
interface because it waits for the "user" to press a key. Because the regex was
too strict, it wouldn't find a match and the call would timeout. So what we have
done is made the regex pattern to be matched more relxed (Line 165).
The other issue with it was how it was building a dictionary containing the
interface information. It was a hodgpe podge because of the timing issues, so
changing the pattern here fixes this (Line 173).

In the nexus driver, sometimes the switch can have "custom" interface names and
when it doesn't match what we expect it crashes, so we changed it so that
continues to run when it encounters such a situation. (Line 159 and 160)
Also, the name of the Port that the switch returns is "Ethernet" with a capital
'E', but we were checking for "ethernet" with a small 'e' so we couldn't find a
match. (Line 164).