URGENT PATCH: Fixes tests of client library
Tests are intermittently failing due to race condition
in the fixture code. Race condition occurs once in a while
when Popen tries to spin a server at the given port before
Popen.terminate of the previous call could release the port.

Fix is to use Popen.wait to ensure that Popen.terminate finishes.
Refactored the fixture code to include this.