Make clientlib tests faster, simpler, and more robust.
The biggest bottleneck was the password hashing -- we really should be
using the null auth backend as much as possible, but we do actually need
to use the password auth backend for some of these tests, and it's
easier to mock out verify the way we have here than to tease these
apart.

We also are no longer spinning up a full web server and networking
daemon. Calling deferred.apply_networking ourselves means we don't have
to worry about race conditions, and can get rid of the calls to sleep.
Doing everything in one process means we can use an in-memory db (for
SQLite, anyway).