Fix a bug in one of the tests.
Fortunately this wasn't failing to catch any bugs it should have, but:
we'd forgotten the 'stuff' argument when making the API call, which was
not what was intended. This patch adds that argument and also adds a
check that the request was successful, which will catch errors like this
in the future.