Make an actual class for switch sessions.
The main reason for this is that inheriting from it means that pylint
won't expect docstrings for the overridden methods, since they're
documented in the superclass.

Ideally, we'd use abstract base classes to get errors earlier, but they
seem to interact badly with sqlalchemy models, which define their own
metaclass.