Improve checks for bad format string usage with logging.
We have our own linter, but pylint checks for this too now. The custom
linter was there because when I wrote it, pylint missed about 1/3 of the
things that it found.

The custom linter previously didn't check for the use of the function
name `warning`, just `warn`, so missed one case. The custom linter has
been modified to cover this case as well.