Clean up redundant auth check pattern.
This generalizes the interface for {require,have}_project_access, so
that passing `None` is valid, and equivalent to {require,have}_admin.
This lets us simplify a whole bunch of stuff like this:

    if project is None:
        require_admin()
    else:
        require_project_access(project)

This also fixes #623, as there is already a check requiring the argument
to be an instance of Project.