Add checks to handle the scenario where there's a failed networking action
* Added this to tests in deferred.py because we have a test switch that can
can raise a switch error.

* This tests that the API removes the failed actions and adds a new one on the
same nic.

* The fixture to create a new_nic was creating a new node every time with the
same name, which lead to failures when using the API. So I have changed it to,
for now, to use new node names every time.

* In the setup of the test we ended up adding 4 actions rather than the
intended 3 which was causing trouble. This was an error that we missed out
when we originally added the test.