Tighten overly-permissive schema on node_register.
previously this would allow any value for 'metadata', while we require a
dictionary. This patch updates the schema to check for the dictionary.

Note that there should be no security problem here; a non-dictionary
would just cause the call to items() to raise an exception, returning a
500 status.

This also makes a change to allow testing the schema: rest_call now
attaches the schema to the function object. See the docstring.