A couple migration related fixes.
This fixes a few missing changes to the Node creation code.

Additionally, this factors out the test-specific logic into its own
migration script. We're having a problem where the changes we're making
to the obmd fields are not being seen by the script that marks those
fields as NOT NULL. This was done in an attempt to fix that, but it
didn't work. It still seems cleaner to me though, so I'm leaving it.