Fix various errors from pycodestyle.
These all cropped up because of changes in the tool. They fall into two
categories:

1. Line-too-long errors in docstrings, which the tool previously missed.
2. Meaningless escape sequences in strings.

Fixing (1) involved minor re-wordings of some docstrings, and changing
the formatting of others.

Only some instances of (2) were fixed, namely those in the test suite
that were windows-style paths, but didn't escape the '\' properly.
Others, which are currently left un-fixed are in regexes in the switch
drivers. These are more delicate to fix, since we have to be careful
not to break the semantics.