Don't get port object in modify port in brocade and dellnos9
Only the console based switches require that to get the old native vlan.
They find the nic_id to which port is connected to, and then get the old_native
from the network_attachment table. old native_vlan is required so that we can
remove that vlan from the list of allowed vlans.

The brocade doesn't need to know the native vlan to remove it.

The dellnos9 implements a method to get the native vlan from the switch and
then removes it. We could implement a similar method for the console based
drivers to do the same.