Rework test http client to handle external redirects.
...and document the requirement that HTTPClient implementations must
follow redirects.

Previously, I was getting a test failure for WIP obmd related tests,
because the client wasn't following a redirect from power_off. Simply
telling the flask test client to follow redirects is insufficient, since
it can't make requests except to the flask app, so the connection to
obmd fails.

However, we also can't just use a regular RequestsHTTPClient, because
that will attempt to make an actual network connection in all cases, and
we aren't actually listening on a socket.

This patch replaces FlaskHTTPClient with a client for testing which uses
either Flask's test client, or the requests library, as appropriate.