Put an upper bound on pytest dependency.
With 3.6.x we're getting an error:

    E ValueError: fixture is being applied more than once to the same
    function

wrt to fail_on_log_warnings in dellnos9.py, if we try to run both it and
the tests in brocade.py in the same run.

We should figure out what's actually going on at some point, but it
seems the pytest devs have broken semver, so we're just going to set
a lower upper bound for now.