Check for pending actions when we actually have one
This was pointed out by @ianballou. In case there are no nics on a node
or there are no networks attached to those nics, then we skip the while
loop that checks for networking actions (since there won't be any).

Boolean `done` is set to True in the beginning, and is only set to False
when we find a nic with a network attached to it.